+++ build.gradle
@@ -146,6 +146,12 @@ dependencies {
     compile group: 'javax.annotation', name: 'javax.annotation-api', version: '1.2-b01'
     compile group: 'com.github.stephenc.eaio-uuid', name: "uuid", version: "3.4.0"
 
+    compile (group: 'com.github.docker-java', name: 'docker-java-transport-okhttp', version: '3.2.1') {
+        exclude group: 'commons-io', module: 'commons-io'
+        exclude group: 'com.fasterxml.jackson'
+        exclude group: 'com.fasterxml.jackson.core'
+    }
+
     compile "org.jgrapht:jgrapht-core:1.3.0"
     implementation 'com.github.LBNL-UCB-STI:helics-wrapper:v2.4.1-2'
