+++ routing-framework/Dockerfile
@@ -9,7 +9,7 @@ RUN apt update && apt --yes install tzdata \
         libcgal-dev libproj-dev zlib1g-dev
 
 RUN git clone https://github.com/LBNL-UCB-STI/routing-framework.git && \
-    cd routing-framework && git submodule update --init && git checkout exp
+    cd routing-framework && git submodule update --init
 
 SHELL ["/bin/bash", "-c"]
 WORKDIR routing-framework/External
@@ -19,8 +19,6 @@ RUN  cd fast-cpp-csv-parser && cp *.h /usr/local/include && cd .. && \
      cd RoutingKit && make && cp -r include lib /usr/local && cd .. && \
      cd stocc && mkdir /usr/local/include/stocc && cp *.cpp *.h $_ && cd .. && \
      cd vectorclass && mkdir /usr/local/include/vectorclass && cp *.h special/* $_ && cd ..
-
-RUN git pull origin exp && echo 1
 
 WORKDIR /routing-framework
