+++ src/main/scala/beam/physsim/jdeqsim/PhysSimReplayer.scala
@@ -5,7 +5,7 @@ import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 
 import beam.agentsim.events.PathTraversalEvent
-import beam.physsim.routingTool.RoutingFrameworkWrapperImpl
+import beam.physsim.cch.RoutingFrameworkWrapperImpl
 import beam.sim.{BeamConfigChangesObservable, BeamHelper}
 import beam.sim.config.BeamConfig
 import beam.utils.{BeamConfigUtils, EventReader}
