+++ src/main/scala/beam/sim/BeamHelper.scala
@@ -1,9 +1,8 @@
 package beam.sim
 
-import java.io.{File, FileOutputStream, FileWriter}
+import java.io.FileOutputStream
 import java.nio.file.{Files, Paths, StandardCopyOption}
 import java.time.ZonedDateTime
-import java.util
 import java.util.Properties
 
 import beam.agentsim.agents.choice.mode.{ModeIncentive, PtFares}
@@ -12,15 +11,9 @@ import beam.agentsim.agents.vehicles.VehicleCategory.MediumDutyPassenger
 import beam.agentsim.agents.vehicles._
 import beam.agentsim.events.handling.BeamEventsHandling
 import beam.agentsim.infrastructure.taz.{H3TAZ, TAZTreeMap}
-import beam.agentsim.infrastructure.charging.ChargingPointType
-import beam.agentsim.infrastructure.charging.ElectricCurrentType.DC
-import beam.agentsim.infrastructure.parking.ParkingType.{Public, Residential, Workplace}
-import beam.agentsim.infrastructure.parking.ParkingZoneFileUtils
-import beam.agentsim.infrastructure.taz.TAZTreeMap
 import beam.analysis.ActivityLocationPlotter
 import beam.analysis.plots.{GraphSurgePricing, RideHailRevenueAnalysis}
 import beam.matsim.{CustomPlansDumpingImpl, MatsimConfigUpdater}
-import beam.physsim.routingTool.{RoutingFrameworkWrapper, RoutingFrameworkWrapperImpl}
 import beam.replanning._
 import beam.replanning.utilitybased.UtilityBasedModeChoice
 import beam.router._
@@ -31,10 +24,8 @@ import beam.scoring.BeamScoringFunctionFactory
 import beam.sim.ArgumentsParser.{Arguments, Worker}
 import beam.sim.common.{GeoUtils, GeoUtilsImpl}
 import beam.sim.config._
-import beam.sim.metrics.BeamStaticMetricsWriter
 import beam.sim.metrics.Metrics._
-import beam.sim.metrics.SimulationMetricCollector.{defaultMetricName, SimulationTime}
-import beam.sim.metrics.{InfluxDbSimulationMetricCollector, SimulationMetricCollector}
+import beam.sim.metrics.{BeamStaticMetricsWriter, InfluxDbSimulationMetricCollector, SimulationMetricCollector}
 import beam.sim.modules.{BeamAgentModule, UtilsModule}
 import beam.sim.population.{PopulationAdjustment, PopulationScaling}
 import beam.utils.BeamVehicleUtils.{readBeamVehicleTypeFile, readFuelTypeFile, readVehiclesFile}
@@ -46,31 +37,27 @@ import beam.utils.{NetworkHelper, _}
 import com.conveyal.r5.transit.TransportNetwork
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.scala.DefaultScalaModule
+import com.github.dockerjava.core.DockerClientBuilder
 import com.google.inject
 import com.typesafe.config.{ConfigFactory, Config => TypesafeConfig}
 import com.typesafe.scalalogging.LazyLogging
 import kamon.Kamon
-import org.matsim.api.core.v01.population.Person
 import org.matsim.api.core.v01.{Id, Scenario}
 import org.matsim.core.api.experimental.events.EventsManager
 import org.matsim.core.config.groups.TravelTimeCalculatorConfigGroup
 import org.matsim.core.config.{Config => MatsimConfig}
 import org.matsim.core.controler._
 import org.matsim.core.controler.corelisteners.{ControlerDefaultCoreListenersModule, EventsHandling, PlansDumping}
-import org.matsim.core.population.PersonUtils
 import org.matsim.core.scenario.{MutableScenario, ScenarioBuilder, ScenarioByInstanceModule, ScenarioUtils}
 import org.matsim.core.trafficmonitoring.TravelTimeCalculator
-import org.matsim.households.{Household, HouseholdImpl}
 import org.matsim.utils.objectattributes.AttributeConverter
-import org.matsim.utils.objectattributes.attributable.Attributes
 import org.matsim.vehicles.Vehicle
 
 import scala.collection.JavaConverters._
 import scala.collection.concurrent.TrieMap
-import scala.collection.{mutable, JavaConverters}
 import scala.collection.mutable.ListBuffer
 import scala.concurrent.Await
-import scala.util.Random
+import scala.util.Try
 
 trait BeamHelper extends LazyLogging {
 
@@ -367,7 +354,16 @@ trait BeamHelper extends LazyLogging {
       else parsedArgs.config.get
     }).withFallback(location).resolve()
 
+    checkDockerIsInstalledForCCHPhysSim(config)
+
     (parsedArgs, config)
+  }
+
+  private def checkDockerIsInstalledForCCHPhysSim(config: TypesafeConfig) = {
+    val physSimType = Try(config.getString("beam.physsim.physSimType")).getOrElse("")
+    if (physSimType == "CCH") {
+      DockerClientBuilder.getInstance().build.versionCmd().exec()
+    }
   }
 
   private def postRunActivity(configLocation: String, config: TypesafeConfig, outputDirectory: String) = {
