+++ test/input/beamville/beam.conf
@@ -1,7 +1,7 @@
 include "../common/akka.conf"
 include "../common/metrics.conf"
 include "../common/matsim.conf"
-
+beam.physsim.physSimType = "CCH"
 ##################################################################
 # Agentsim
 ##################################################################
